chore/localization-and-word-fixes * move localization files to a new directory and update imports * update translations
